Added monthwise statistics with basic d3 pie chart

+++ src/main/java/com/ridetrends/bean/UberMonthlyStats.java
@@ -3,40 +3,40 @@ package com.ridetrends.bean;
 import java.util.Map;
 
 public class UberMonthlyStats {
-	private Map<Integer, Long> monthWiseRidesCount;
-	private Map<Integer, Double> monthWiseTotalDistanceCovered;
-	private Map<Integer, Double> monthWiseTotalRideTimes;
-	private Map<Integer, Double> monthWiseTotalWaitingTimes;
+	private Map<Long, Long> monthWiseRidesCount;
+	private Map<Long, Double> monthWiseTotalDistanceCovered;
+	private Map<Long, Double> monthWiseTotalRideTimes;
+	private Map<Long, Double> monthWiseTotalWaitingTimes;
 
-	public Map<Integer, Long> getMonthWiseRidesCount() {
+	public Map<Long, Long> getMonthWiseRidesCount() {
 		return monthWiseRidesCount;
 	}
 
-	public void setMonthWiseRidesCount(Map<Integer, Long> monthWiseRidesCount) {
+	public void setMonthWiseRidesCount(Map<Long, Long> monthWiseRidesCount) {
 		this.monthWiseRidesCount = monthWiseRidesCount;
 	}
 
-	public Map<Integer, Double> getMonthWiseTotalDistanceCovered() {
+	public Map<Long, Double> getMonthWiseTotalDistanceCovered() {
 		return monthWiseTotalDistanceCovered;
 	}
 
-	public void setMonthWiseTotalDistanceCovered(Map<Integer, Double> monthWiseTotalDistanceCovered) {
+	public void setMonthWiseTotalDistanceCovered(Map<Long, Double> monthWiseTotalDistanceCovered) {
 		this.monthWiseTotalDistanceCovered = monthWiseTotalDistanceCovered;
 	}
 
-	public Map<Integer, Double> getMonthWiseTotalRideTimes() {
+	public Map<Long, Double> getMonthWiseTotalRideTimes() {
 		return monthWiseTotalRideTimes;
 	}
 
-	public void setMonthWiseTotalRideTimes(Map<Integer, Double> monthWiseTotalRideTimes) {
+	public void setMonthWiseTotalRideTimes(Map<Long, Double> monthWiseTotalRideTimes) {
 		this.monthWiseTotalRideTimes = monthWiseTotalRideTimes;
 	}
 
-	public Map<Integer, Double> getMonthWiseTotalWaitingTimes() {
+	public Map<Long, Double> getMonthWiseTotalWaitingTimes() {
 		return monthWiseTotalWaitingTimes;
 	}
 
-	public void setMonthWiseTotalWaitingTimes(Map<Integer, Double> monthWiseTotalWaitingTimes) {
+	public void setMonthWiseTotalWaitingTimes(Map<Long, Double> monthWiseTotalWaitingTimes) {
 		this.monthWiseTotalWaitingTimes = monthWiseTotalWaitingTimes;
 	}
